requirements.txt
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dlib==19.24.0
flake8==4.0.1
Flask==2.1.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
numpy==1.22.4
opencv-python==4.6.0.66
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
requests==2.28.0
scipy==1.8.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
